gateway-api: extract listener status handling
[ upstream commit d7dd695adaf7095ce822f6406065ad164b95a582 ]

Move Gateway and ListenerSet listener status handling into
a dedicated `ListenerStatusManager`.

This refactor moves shared listener status behavior out of
`gatewayReconciler`. The reconciler now owns less listener-specific
logic and calls the manager directly for ListenerSet status.
